+++ core/src/tui_core/tui_style/mod.rs
@@ -16,16 +16,16 @@
  */
 
 // Attach sources.
+pub mod crossterm_color_converter;
 pub mod hex_color_parser;
 pub mod tui_color;
 pub mod tui_style_impl;
 pub mod tui_style_lite;
 pub mod tui_stylesheet;
-pub mod crossterm_color_converter;
 
 // Re-export.
+pub use crossterm_color_converter::*;
 pub use hex_color_parser::*;
 pub use tui_color::*;
 pub use tui_style_impl::*;
 pub use tui_stylesheet::*;
-pub use crossterm_color_converter::*;