+++ tui/src/tui/terminal_window/shared_global_data.rs
@@ -19,22 +19,18 @@ use std::fmt::{Debug, Formatter};
 
 use r3bl_core::{ok,
                 sizing::TelemetryReportLineStorage,
+                spinner_impl,
                 ChUnit,
                 CommonResult,
                 InlineString,
                 OutputDevice,
                 Size,
+                SpinnerStyle,
                 TelemetryHudReport};
 use tokio::sync::mpsc::Sender;
 
 use super::TerminalWindowMainThreadSignal;
-use r3bl_core::{spinner_impl,SpinnerStyle,};
-use crate::{
-            OffscreenBuffer,
-            OffscreenBufferPool,
-
-            DEBUG_TUI_COMPOSITOR,
-            DEBUG_TUI_MOD};
+use crate::{OffscreenBuffer, OffscreenBufferPool, DEBUG_TUI_COMPOSITOR, DEBUG_TUI_MOD};
 
 /// This is a global data structure that holds state for the entire application
 /// [crate::App] and the terminal window [crate::TerminalWindow] itself.